// src/app/modules/celebrities-details/interfaces/celebrity-crew.interface.ts
export interface CelebrityCrew {
  adult: boolean;
  backdrop_path: null;
  credit_id: string;
  department: string;
  genre_ids: [];
  id: number;
  job: string;
  media_type: string;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: null;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}